fixed a horrible bug which was causing wrong info in correct caption
The correct caption was showing very wierd values.
After debugging, I found out it has wrong values and the culpirt was batch_first parameter which is passed to pad_sequence in caps_collate class
As a consequence, the shape of the captions has changed and all Transposing with (.T) has to be removed
Accuracy is still nan, and needs to be fixed
Another fix which I tried in here is to set the accurcay to zero when an error occurs.
I left some comments on purpose to come back again to this.